isolate draft functions
